x1plusd/expansion: refactor FTDI-specific logic out into ft2232/ to prepare for rp2040/ boards
