+++ images/cfw/opt/x1plus/lib/python/x1plus/services/x1plusd/expansion/device.py
@@ -1,0 +1,14 @@
+import abc
+
+class ExpansionDevice(abc.ABC):
+    revision = None
+    serial = None
+    nports = 0
+    
+    @abc.abstractmethod
+    def detect_eeprom(self, port):
+        pass
+    
+    @abc.abstractmethod
+    def reset(self):
+        pass
