+++ images/cfw/opt/x1plus/lib/python/x1plus/services/x1plusd/expansion/ft2232/detect_eeprom.py
