+++ images/cfw/opt/x1plus/lib/python/x1plus/services/x1plusd/expansion/manager.py
@@ -9,7 +9,7 @@ import pyftdi.ftdi
 
 from ..dbus import *
 
-from .detect_eeprom import detect_eeprom
+from .ft2232 import FtdiExpansionDevice
 
 # workaround for missing ldconfig
 def find_library(lib):
@@ -21,60 +21,20 @@ usb.backend.libusb1.get_backend(find_library=find_library)
 
 logger = logging.getLogger(__name__)
 
-ExpansionDevice = namedtuple('ExpansionDevice', [ 'revision', 'serial', 'ftdidev' ])
-
-def _detect_x1p_002_b01():
-    """
-    Looks for a X1P-002-B01.
-    
-    If it finds one, returns an ExpansionDevice.
-    """
-    
-    lan9514_eth = usb.core.find(idVendor = 0x0424, idProduct = 0xec00)
-    if not lan9514_eth:
-        return None
-    
-    if lan9514_eth.product == 'Expansion Board X1P-002-B01':
-        revision = lan9514_eth.product.split(' ')[2]
-        serial = lan9514_eth.serial_number
-    else:
-        # Maybe it hasn't been serialized; I guess we will limp along in the
-        # mean time, and hope that we can find a sibling FTDI.
-        logger.warning("found a LAN9514, but it does not appear to be an X1P-002-B01?")
-        revision = 'Unknown'
-        serial = 'Unknown'
-    
-    # Look for a sibling FTDI device.
-    ftdidev = usb.core.find(custom_match = lambda d: d.parent == lan9514_eth.parent, idVendor = 0x0403)
-    if not ftdidev:
-        logger.warning("found a LAN9514, but no FTDI sibling")
-        return None
-    
-    # At least on X1P-002-B01, the FT2232 seems to sometimes get confused
-    # about frequency when reopened.  Resetting it seems to make it happier. 
-    # XXX: what causes this -- would a driver reload trigger this too?
-    ftdidev.reset()
-    
-    return ExpansionDevice(revision = revision, serial = serial, ftdidev = ftdidev)
-
 EXPANSION_INTERFACE = "x1plus.expansion"
 EXPANSION_PATH = "/x1plus/expansion"
 
 class ExpansionManager(X1PlusDBusService):
-    DRIVERS = {}
-
     def __init__(self, daemon, **kwargs):
         self.daemon = daemon
 
         self.eeproms = {}
         self.drivers = {}
-        self.ftdi_nports = 0
-        self.ftdi_path = None
         self.last_configs = {}
 
         # We only have to look for an expansion board on boot, since it
         # can't be hot-installed.
-        self.expansion = _detect_x1p_002_b01()
+        self.expansion = FtdiExpansionDevice.detect()
         if not self.expansion:
             logger.info("no X1Plus expansion board detected")
             super().__init__(
@@ -84,17 +44,10 @@ class ExpansionManager(X1PlusDBusService):
         
         logger.info(f"found X1Plus expansion board serial {self.expansion.serial}")
         
-        self.ftdi_path = f"ftdi://::{self.expansion.ftdidev.bus:x}:{self.expansion.ftdidev.address:x}/"
-        if self.expansion.ftdidev.idProduct == 0x6010: # FT2232H
-            self.ftdi_nports = 2
-        else:
-            self.ftdi_nports = 2
-            logger.warning(f"FTDI product ID {self.expansion.ftdidev.idProduct:x} unrecognized")
-
-        for port in range(self.ftdi_nports):
+        for port in range(self.expansion.nports):
             port_name = f"port_{chr(0x61 + port)}"
             self.eeproms[port_name] = None
-            eeprom = detect_eeprom(f"{self.ftdi_path}{port + 1}")
+            eeprom = self.expansion.detect_eeprom(port)
             if eeprom:
                 try:
                     model, revision = eeprom[:16].decode().strip().rsplit('-', 1)
@@ -104,7 +57,7 @@ class ExpansionManager(X1PlusDBusService):
                 except:
                     logger.error(f"error decoding EEPROM contents {eeprom} on {port_name}")
         
-        for port in range(self.ftdi_nports):
+        for port in range(self.nports):
             self.daemon.settings.on(f"expansion.port_{chr(0x61 + port)}", lambda: self._update_drivers())
 
         self.last_configs = {}
@@ -124,7 +77,7 @@ class ExpansionManager(X1PlusDBusService):
         # Workaround https://github.com/eblot/pyftdi/issues/261 by resetting
         # all drivers on the FTDI every time.
         did_change = False
-        for port in range(self.ftdi_nports):
+        for port in range(self.expansion.nports):
             port_name = f"port_{chr(0x61 + port)}"
             config = self.daemon.settings.get(f"expansion.{port_name}", None)
             if self.daemon.settings.get(f"expansion.{port_name}", None) != self.last_configs.get(port_name, None):
@@ -133,7 +86,7 @@ class ExpansionManager(X1PlusDBusService):
         
         if did_change:
             # shut down all ports...
-            for port in range(self.ftdi_nports):
+            for port in range(self.expansion.nports):
                 port_name = f"port_{chr(0x61 + port)}"
                 if port_name in self.drivers:
                     self.drivers[port_name].disconnect()
@@ -143,9 +96,9 @@ class ExpansionManager(X1PlusDBusService):
                     del self.last_configs[port_name]
             
             # reset the FTDI ...
-            self.expansion.ftdidev.reset()
+            self.expansion.reset()
 
-        for port in range(self.ftdi_nports):
+        for port in range(self.expansion.nports):
             port_name = f"port_{chr(0x61 + port)}"
             config = self.daemon.settings.get(f"expansion.{port_name}", None)
             if not config:
@@ -173,12 +126,12 @@ class ExpansionManager(X1PlusDBusService):
             driver = ckey.pop()
             subconfig = config[driver]
             
-            if driver not in self.DRIVERS:
-                logger.error(f"{port_name} is assigned driver {driver}, which is not registered")
+            if driver not in self.expansion.DRIVERS:
+                logger.error(f"{port_name} is assigned driver {driver}, which is not valid for this Expander")
                 continue
             
             try:
-                self.drivers[port_name] = self.DRIVERS[driver](ftdi_path = f"{self.ftdi_path}{port + 1}", port_name = port_name, config = subconfig, daemon = self.daemon)
+                self.drivers[port_name] = self.expansion.DRIVERS[driver](ftdi_path = f"{self.expansion.ftdi_path}{port + 1}", port_name = port_name, config = subconfig, daemon = self.daemon)
                 self.last_configs[port_name] = config
             except Exception as e:
                 logger.error(f"{port_name} driver {driver} initialization failed: {e.__class__.__name__}: {e}")
